# backend/microservices/admin-service/swagger.json
{
    "swagger": "2.0",
     "info": {
       "description": "This is a sample-service server",
       "version": "1.0.0",
       "title": "Sample Service"
     },
     "host": "localhost:3000",
     "basePath": "/api/v1/sample",
     "paths": {
       "/home": {
         "get": {
           "tags": [
             "Sample API"
           ],
           "summary": " /home API",
           "description": "",
           "produces": [
             "application/json"
           ],
           "responses": {
             "200": {
               "description": "Success"
             },
             "404": {
               "description": "Not Found"
             },
             "500": {
               "description": "Internal Error"
             }
           }
         }
       },
       "/users": {
         "post": {
           "tags": [
             "Sample API"
           ],
           "summary": "Creates user",
           "description": "",
           "operationId": "createUser",
           "consumes": [
             "application/json"
           ],
           "produces": [
             "application/json"
           ],
           "parameters": [
             {
               "in": "body",
               "name": "body",
               "description": "User object",
               "required": true,
               "schema": {
                 "type": "object",
                 "items": {
                   "$ref": "#/definitions/User"
                 }
               }
             }
           ],
           "responses": {
             "200": {
               "description": "Success"
             },
             "404": {
               "description": "Not Found"
             },
             "500": {
               "description": "Internal Error"
             }
           }
         }
       }
     },
     "definitions": {
       "User": {
         "type": "object",
         "properties": {
           "id": {
             "type": "integer",
             "format": "int64"
           },
           "name": {
             "type": "string"
           },
           "email":{
             "type":"string"
           }
         }
       }
     }
   }